backend/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
ipykernel==6.16.1
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.5
jupyter_core==4.12.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.3
postgres==4.0
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2-binary==2.9.5
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==24.0.1
six==1.16.0
SQLAlchemy==2.0.2
sqlalchemy-utils==0.40.0
stack-data==0.6.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.12.1
